Use retry in S3Streamer
