+++ multiurl/__init__.py
@@ -8,12 +8,13 @@
 #
 
 
-from .downloader import Downloader, download, robust
+from .downloader import Downloader, download
+from .retry import robust
 
 __version__ = "0.3.8dev0"
 
 __all__ = [
     "download",
     "Downloader",
-    "robust",
+    "robust"
 ]
